Split haptic plug and haptic socket builders
