+++ com.vrcfury.vrcfury/Editor/VF/Feature/BakeHapticPlugsBuilder.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 37390e6dbb5f48ecaff095183e77bff5
+timeCreated: 1693301913
