+++ com.vrcfury.vrcfury/Editor/VF/Feature/BakeHapticSocketsBuilder.cs.meta
