+++ com.vrcfury.vrcfury/Editor/VF/Feature/FixTouchingContactsService.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 4b912c4836f04620a03a229919bd8a8d
+timeCreated: 1693302273
